// src/app/models/providers.model.ts
export interface Providers {
    id: string; // Identificador único del proveedor
    nombre: string; // Nombre del proveedor
    telefono: string; // Número de teléfono de contacto
    email: string; // Correo electrónico del proveedor
    notas: string; // Campo para notas adicionales sobre el proveedor
    fecha: string; // Fecha de registro del proveedor en el sistema
    activo: string; // Estado activo/inactivo del proveedor
    img: string; // Imagen del proveedor
  }